intelhexfileio:	data type for adr, len of intelhex data blocks changed to quint32

+++ wm3000id.cpp
@@ -1093,8 +1093,8 @@ const char* cWM3000iServer::mControlerMemUpdate(bl_cmdcode blwriteCmd,char* s)
 		    for (i = 0; i < 4; i++)
 			dest[i ^ 1] = blInput[pos+1+i]; // little endian ... big endian
 		    dest[i] = blInput[pos+1+i];
-                    int MemAdress = 0;
-		    ulong MemOffset;
+                    quint32 MemAdress = 0;
+                    quint32 MemOffset;
 		    QByteArray MemByteArray;
 		    IntelHexData.GetMemoryBlock( BootloaderInfo.MemPageSize, MemAdress, MemByteArray, MemOffset);
 		    while ( (MemByteArray.count()) && !(execError) ) // solange wir noch daten aus hexfile bekommen
